質問の種類によって実装が切り替わるように
