+++ cmd/picha/main.go
@@ -1,10 +1,8 @@
 package main
 
 import (
-	"bufio"
 	"fmt"
 	"os"
-	"path/filepath"
 
 	"github.com/AlecAivazis/survey/v2"
 	picha "github.com/ddddddO/pipe-chatgpt"
@@ -39,76 +37,9 @@ func do() error {
 		return err
 	}
 
-	// TODO: ここは、なにかinterfaceで定義して、その実装を呼び出す、に変更したいかも
-	switch answerType {
-	case "テキスト":
-		for {
-			answerText := ""
-			if err := survey.AskOne(
-				&survey.Input{
-					Message: "聞きたいことは？",
-				}, &answerText, survey.WithValidator(survey.Required)); err != nil {
-				return err
-			}
-			if err := gptClient.RequestToDavinci(answerText); err != nil {
-				return err
-			}
-		}
-	case "テキストファイル":
-		answerPath := ""
-		if err := survey.AskOne(
-			&survey.Input{
-				Message: "そのファイルパスを入力してください",
-			}, &answerPath, survey.WithValidator(survey.Required)); err != nil {
-			return err
-		}
-		path, err := filepath.Abs(answerPath)
-		if err != nil {
-			return err
-		}
-		if !isExist(path) {
-			return fmt.Errorf("no exist file: %s", answerPath)
-		}
-
-		answerProcessingFile := ""
-		if err := survey.AskOne(
-			&survey.Input{
-				Message: "このファイルをどうしたいですか？",
-			}, &answerProcessingFile, survey.WithValidator(survey.Required)); err != nil {
-			return err
-		}
-
-		f, err := os.Open(path)
-		if err != nil {
-			return err
-		}
-		defer f.Close()
-
-		// TODO: 一旦、jsonとかcsvとか考えずに普通のテキストファイルとして作る
-		sc := bufio.NewScanner(f)
-		in := "「"
-		for sc.Scan() {
-			in += sc.Text()
-		}
-		if err := sc.Err(); err != nil {
-			return err
-		}
-		in += "」"
-		in += answerProcessingFile
-
-		return gptClient.RequestToDavinci(in)
-	case "音声":
-		// TODO:
-	default:
-		return fmt.Errorf("未定義の種類です")
+	responser, err := picha.ResponserFactory(answerType, gptClient)
+	if err != nil {
+		return err
 	}
-
-	return nil
-}
-
-func isExist(path string) bool {
-	if _, err := os.Stat(path); !os.IsNotExist(err) {
-		return true
-	}
-	return false
+	return responser.Run()
 }
